feat(bedrock)!: move BedrockOpenAIOkHttpClient to com.openai.bedrock
`openai-java-bedrock` shipped `BedrockOpenAIOkHttpClient` in the
`com.openai.client.okhttp` package, which is owned by
`openai-java-client-okhttp`. Because the Bedrock artifact declares a
compile-scope dependency on the OkHttp artifact, every consumer has both
JARs on the path, so the two always split that package.

On the class path this is harmless, but it makes the Bedrock artifact
unusable on the module path:

    java.lang.module.ResolutionException: Modules com.openai.client.okhttp
    and com.openai.bedrock export package com.openai.client.okhttp to
    module com.openai.core

Relocating the class is the only fix: a deprecated forwarding class left
behind in the old package would keep the package present in both JARs and
preserve the split. The class only ever used public `OpenAIOkHttpClient`
builder methods, so its original placement was cosmetic rather than
load-bearing.

BREAKING CHANGE: `BedrockOpenAIOkHttpClient` moved from
`com.openai.client.okhttp` to `com.openai.bedrock`. Update imports to
`import com.openai.bedrock.BedrockOpenAIOkHttpClient;`.
