+++ openai-java-example/src/main/java/com/openai/example/BedrockResponsesExample.java
@@ -1,7 +1,7 @@
 package com.openai.example;
 
+import com.openai.bedrock.BedrockOpenAIOkHttpClient;
 import com.openai.client.OpenAIClient;
-import com.openai.client.okhttp.BedrockOpenAIOkHttpClient;
 import com.openai.models.responses.ResponseCreateParams;
 
 public final class BedrockResponsesExample {
